+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## v1.3.2
+- [#34](https://github.com/lastorel/pytion/issues/34): `status` Property type added
+
 ## v1.3.1
 - [#32](https://github.com/lastorel/pytion/issues/32): Rollback of Page retrieving with properties
 - [#35](https://github.com/lastorel/pytion/issues/35): Fix PropertyValue with rollup type